% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{antidiabetic}
\alias{antidiabetic}
\title{Phung et al. (2010)'s network meta-analysis data}
\format{An arm-based dataset with 20 studies}
\usage{
data(antidiabetic)
}
\description{
A network meta-analysis dataset for treatments of type-2 diabetes from Chaimani and Salanti (2015).
\itemize{
  \item \code{id}: Study ID
  \item \code{t}: Treatment (Placebo, AGI, DPP-4 inhibitor, Glinine, GLP-1 analog, Sulfonylurea, Thiazolidinedione)
  \item \code{y}: Mean of the change in HbA1c
  \item \code{sd}: Standard deviation of the change in HbA1c
  \item \code{n}: Sample size
}
}
\references{
Chaimani, A. and Salanti, G. (2015). Visualizing assumptions and results in network meta-analysis: the network graphs package. \emph{Stata Journal.} \strong{15}: 905-920. 

Phung, O. J., Scholle, J. M., Talwar, M. and Coleman, C. I. (2010). Effect of noninsulin antidiabetic drugs added to metformin therapy on glycemic control, weight gain, and hypoglycemia in type 2 diabetes. 
\emph{JAMA.} \strong{303}: 1410-1418.
}
\keyword{datasets}
