% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surroNMA.R
\name{augment_network_with_SI}
\alias{augment_network_with_SI}
\title{Apply a trained SI to augment the network}
\usage{
augment_network_with_SI(net, si_model)
}
\arguments{
\item{net}{Network object from surro_network}

\item{si_model}{Trained surrogate index model}
}
\value{
Augmented network object
}
\description{
Apply a trained SI to augment the network
}
