% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binData}
\alias{binData}
\title{Binary data}
\format{
A data frame with 5 variables:
\describe{
\item{study}{Study name}
\item{treatment}{Reference treatment name}
\item{n}{Total number}
\item{r}{Number of sucesses}
\item{prob}{Percentage successes}
}
}
\usage{
binData
}
\description{
A dataset containing the pair comparisons for studies.
}
\keyword{datasets}
