% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surroNMA.R
\name{compute_STE}
\alias{compute_STE}
\title{Compute Surrogate Threshold Effect (STE)}
\usage{
compute_STE(draws_alpha, draws_beta, threshold_T = 0)
}
\arguments{
\item{draws_alpha}{Draws from posterior distribution of alpha}

\item{draws_beta}{Draws from posterior distribution of beta}

\item{threshold_T}{Threshold value for true endpoint (default 0)}
}
\value{
List with STE values and summary statistics
}
\description{
Compute Surrogate Threshold Effect (STE)
}
