% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contsData}
\alias{contsData}
\title{Continuous data}
\format{
A data frame with 6 variables:
\describe{
\item{study}{Study name}
\item{treatment}{Reference treatment name}
\item{dur}{Duration}
\item{se}{Standard error}
\item{n}{Total number}
\item{y}{Outcome measurement}
}
}
\usage{
contsData
}
\description{
A dataset containing the pair comparisons for studies.
}
\keyword{datasets}
