% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countData}
\alias{countData}
\title{Counts data}
\format{
A data frame with 4 variables:
\describe{
\item{study}{Study name}
\item{treatment}{Reference treatment name}
\item{E}{Total time at risk}
\item{r}{Number of counts}
}
}
\usage{
countData
}
\description{
A dataset containing the pair comparisons for studies.
}
\keyword{datasets}
