% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bugs_code.R
\name{create_bugs_code}
\alias{create_bugs_code}
\title{Create BUGS code from component parts}
\usage{
create_bugs_code(random, dat)
}
\arguments{
\item{random}{Random effect model? Logical}

\item{dat}{NMA object}
}
\description{
Create BUGS code from component parts
}
