% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolders.R
\name{createFolders}
\alias{createFolders}
\title{create output folders}
\usage{
createFolders(folder = "output", fileSep = "/", ...)
}
\arguments{
\item{folder}{}

\item{fileSep}{}

\item{...}{}
}
\value{

}
\description{
create output folders
}
