\docType{data}
\name{exdataMD}
\alias{exdataMD}
\title{Example data of summary statistics from Phung et al. (2010)'s network meta-analysis data}
\format{A data frame for network meta-analysis with 3 trials.}
\usage{
data(exdataMD)
}
\description{
Summary statistics for 3 trials of the network meta-analysis in Phung et al. (2010).
\itemize{
  \item \code{id}: ID variable of studies.
  \item \code{treat1}: Treatment 1.
  \item \code{treat2}: Treatment 2.
  \item \code{MD}: Mean difference estimate.
  \item \code{seMD}: Standard error estimate of the mean difference estimator.
  \item \code{n1}: Sample size 1.
  \item \code{n2}: Sample size 2.
}
}
\references{
Chaimani, A. and Salanti, G. (2015). Visualizing assumptions and results in network meta-analysis: the network graphs package. \emph{Stata Journal.} \strong{15}: 905-920. 

Phung, O. J., Scholle, J. M., Talwar, M. and Coleman, C. I. (2010). Effect of noninsulin antidiabetic drugs added to metformin therapy on glycemic control, weight gain, and hypoglycemia in type 2 diabetes. 
\emph{JAMA.} \strong{303}: 1410-1418.
}

\keyword{datasets}
