\docType{data}
\name{exdataP}
\alias{exdataP}
\title{Example data of arm-specific survival probability estimates for a network meta-analysis}
\format{A data frame for network meta-analysis with 5 trials.}
\usage{
data(exdataP)
}
\description{
Summary statistics for 5 trials of the network meta-analysis.
\itemize{
  \item \code{study}: ID variable of studies.
  \item \code{trt}: Treatment.
  \item \code{y}: Survival probability estimate.
  \item \code{se}: Standard error estimate of the survival probability.
}
}

\keyword{datasets}
