\docType{data}
\name{exdataRR}
\alias{exdataRR}
\title{Example data of summary statistics from Sciarretta et al. (2011)'s network meta-analysis data}
\format{A data frame for network meta-analysis with 3 trials.}
\usage{
data(exdataRR)
}
\description{
Summary statistics for 3 trials of the network meta-analysis in Sciarretta et al. (2011).
\itemize{
  \item \code{id}: ID variable of studies.
  \item \code{treat1}: Treatment 1.
  \item \code{treat2}: Treatment 2.
  \item \code{logRR}: Log risk-ratio estimate.
  \item \code{SE}: Standard error estimate of the log risk-ratio estimator.
  \item \code{n1}: Sample size 1.
  \item \code{n2}: Sample size 2.
}
}
\references{
Sciarretta, S., Palano, F., Tocci, G., Baldini, R., and Volpe, M. (2011). 
Antihypertensive treatment and development of heart failure in hypertension: a Bayesian network meta-analysis of studies in patients with hypertension and high cardiovascular risk. 
\emph{Archives of Internal Medicine} 
\strong{171}: 384-394.
}

\keyword{datasets}
