% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSamps.R
\name{genSamps}
\alias{genSamps}
\title{WinnewBugs Function}
\usage{
genSamps(
  bugsData,
  n.iter = N.SIMS + N.BURNIN,
  inits,
  n.burnin = N.BURNIN,
  newBugs.file,
  parameters.to.save,
  winDebug = PAUSE,
  folder
)
}
\description{
WinnewBugs Function
}
