% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{heartfailure}
\alias{heartfailure}
\title{Sciarretta et al. (2011)'s network meta-analysis data}
\format{An arm-based dataset with 26 studies}
\usage{
data(heartfailure)
}
\description{
\itemize{
  \item \code{study}: Study ID
  \item \code{trial}: Trial name
  \item \code{trt}: Treatment (AB (Alpha blocker), ACE (ACE inhibitor), ARB, BB (Beta blocker), CCB (Calcium-channel blocker), CT (conventional treatments), Diuretic (DD), Placebo)
  \item \code{n}: Sample size
  \item \code{d}: Number of events (occurrence of heart failure)
}
}
\references{
Sciarretta, S., Palano, F., Tocci, G., Baldini, R., and Volpe, M. (2011). 
Antihypertensive treatment and development of heart failure in hypertension: a Bayesian network meta-analysis of studies in patients with hypertension and high cardiovascular risk. 
\emph{Archives of Internal Medicine} 
\strong{171}: 384-394.
}
\keyword{datasets}
