% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{load_input_data}
\alias{load_input_data}
\title{load_input_data}
\usage{
load_input_data(is_med, is_bin, analysis_type, folder = here::here("raw_data"))
}
\arguments{
\item{is_med}{logical}

\item{is_bin}{logical}

\item{folder}{}
}
\value{

}
\description{
load_input_data
}
