% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bugs_filename.R
\name{make_bugs_filename}
\alias{make_bugs_filename}
\title{Make BUGS filename}
\usage{
make_bugs_filename(random, dat)
}
\arguments{
\item{random}{Type of model, RE or FE; logical}

\item{dat}{List of data frames. Possible from subData/Med/Bin}
}
\value{
string
}
\description{
Make BUGS filename
}
