% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_bugs_filename.R
\name{name_bugs_filename}
\alias{name_bugs_filename}
\title{name_bugs_filename}
\usage{
name_bugs_filename(random, dat)
}
\arguments{
\item{random}{logical}

\item{dat}{list}
}
\value{
string
}
\description{
name_bugs_filename
}
