% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_NMA.R
\name{new_NMA}
\alias{new_NMA}
\title{NMA constructor}
\usage{
new_NMA(
  subData,
  subDataMed = NA,
  subDataBin = NA,
  bugs_params = NA,
  is_random = TRUE,
  hyperparams = list(),
  refTx = NA,
  effectParam,
  modelParams,
  label,
  endpoint
)
}
\arguments{
\item{subData}{Main input data frame. Mandatory}

\item{subDataMed}{Median time input data frame. Optional}

\item{subDataBin}{Binary data input data frame. Optional}

\item{bugs_params}{List of BUGS parameters. Optional}

\item{is_random}{Random effects model? Logical}

\item{hyperparams}{List of hyperparameters}

\item{refTx}{Reference treatment; string}

\item{effectParam}{Effect parameter}

\item{modelParams}{Model parameter}

\item{label}{Label}

\item{endpoint}{End point name; string}
}
\value{

}
\description{
Create an \code{nma} class object to use in an analysis.
}
\seealso{
\code{\link{NMA_run}}, \code{\link{NMA_update}}
}
