% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_NMA_dir.R
\name{new_NMA_dir}
\alias{new_NMA_dir}
\title{Load all input data from a folder}
\usage{
new_NMA_dir(data_dir = ".")
}
\arguments{
\item{data_dir}{folder location; string}
}
\value{
\code{nma} object
}
\description{
Load all input data from a folder
}
