% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_nma_output_dir.R
\name{new_nma_output_dir}
\alias{new_nma_output_dir}
\title{Create output folders}
\usage{
new_nma_output_dir(nma, folder = "output", ...)
}
\arguments{
\item{nma}{Object of class \code{nma}}

\item{folder}{Output folder. Text string}

\item{...}{Additional arguments}
}
\value{

}
\description{
Create output folders
}
\seealso{
new_NMA
}
