% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA.R
\name{NMA}
\alias{NMA}
\title{Network meta-analysis}
\usage{
NMA(
  dat,
  bugs_params,
  bugs_fn,
  effectParam = NA,
  modelParams = NA,
  output_dir = "output",
  label,
  endpoint,
  preRefTx = NA,
  random = FALSE,
  RUN = TRUE,
  fileSep = "\\\\"
)
}
\arguments{
\item{dat}{result of setupData()}

\item{bugs_fn}{BUGS function call}

\item{effectParam}{}

\item{modelParams}{}

\item{label}{}

\item{endpoint}{}

\item{preRefTx}{what is this?}

\item{random}{}

\item{fileSep}{File separator; string}
}
\value{

}
\description{
Network meta-analysis
}
