% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_outputs.R
\name{nma_outputs}
\alias{nma_outputs}
\title{All output functions}
\usage{
nma_outputs(
  nma,
  res_bugs,
  effectParam,
  labels,
  endpoint,
  output_dir = "output",
  fileSep = "/"
)
}
\description{
All output functions
}
