% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_run.R
\name{NMA_run.nma}
\alias{NMA_run.nma}
\title{Network meta-analysis}
\usage{
\method{NMA_run}{nma}(nma, fileSep = "/", output_dir = "output", save = TRUE)
}
\arguments{
\item{nma}{object of class nma}

\item{fileSep}{File separator; string}

\item{output_dir}{}

\item{save}{save bugs output to file; logical}
}
\value{

}
\description{
Network meta-analysis
}
