% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_run.R
\name{NMA_run}
\alias{NMA_run}
\alias{NMA_run.nma}
\title{Run network meta-analysis}
\usage{
NMA_run(nma, folder = "output", save = TRUE)

\method{NMA_run}{nma}(nma, folder = "output", save = TRUE)
}
\arguments{
\item{nma}{Object of class \code{nma}}

\item{folder}{Output folder name; string}

\item{save}{Save BUGS output to file? Logical}
}
\value{
\code{res_bugs}


}
\description{
BUGS code is taken from DSU reports and papers.
}
\details{
The survival analysis data used is from the \insertCite{Woods2010}{NMA} model.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{new_NMA}}, \code{\link{NMA_update}}
}
