% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_update.R
\name{NMA_update}
\alias{NMA_update}
\alias{NMA_update.nma}
\title{NMA_update}
\usage{
NMA_update(nma, ...)

\method{NMA_update}{nma}(nma, ...)
}
\arguments{
\item{nma}{}

\item{...}{}
}
\value{



}
\description{
NMA_update
}
