% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseTable.R
\name{pairwiseTable}
\alias{pairwiseTable}
\title{Pairwise table}
\usage{
pairwiseTable(
  nma,
  res_bugs,
  folder = "output",
  probs = c(0.025, 0.975),
  label = "",
  save = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
table
}
\description{
Pairwise table
}
