% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{placeholder-functions}
\alias{placeholder-functions}
\title{Placeholder Functions}
\arguments{
\item{K}{Number of treatments}

\item{J}{Number of studies}

\item{per_study}{Studies per treatment}

\item{seed}{Random seed}
}
\value{
NULL (placeholder)
}
\description{
These functions are placeholders for future development
}
\keyword{internal}
