% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork.default}
\alias{plotNetwork.default}
\title{plotNetwork}
\usage{
\method{plotNetwork}{default}(dat, usecurve = FALSE, ...)
}
\arguments{
\item{dat}{}

\item{usecurve}{}

\item{...}{}
}
\value{

}
\description{
plotNetwork
}
