% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{plotNetwork}
\usage{
plotNetwork(dat, usecurve = FALSE, ...)
}
\arguments{
\item{dat}{}

\item{usecurve}{}

\item{...}{}
}
\value{

}
\description{
plotNetwork
}
