% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_and_tables.R
\name{plots_and_table}
\alias{plots_and_table}
\title{plots_and_table}
\usage{
plots_and_table(dat, effectParam, labels, fileSep = "/")
}
\arguments{
\item{dat}{}

\item{effectParam}{}

\item{labels}{}

\item{fileSep}{}
}
\value{

}
\description{
plots_and_table
}
