% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_and_tables.R
\name{plots_and_tables}
\alias{plots_and_tables}
\title{plots_and_tables}
\usage{
plots_and_tables(
  dat,
  res_bugs,
  effectParam,
  labels,
  endpoint = NULL,
  folder = "output",
  fileSep = "/"
)
}
\arguments{
\item{dat}{}

\item{effectParam}{}

\item{labels}{}

\item{folder}{}

\item{fileSep}{}
}
\value{

}
\description{
plots_and_tables
}
