% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bin_data.R
\name{prep_bin_data}
\alias{prep_bin_data}
\title{prep_bin_data}
\usage{
prep_bin_data(binData, refTx = NA, preRefTx = NA)
}
\arguments{
\item{binData}{}

\item{refTx}{}

\item{preRefTx}{}
}
\value{

}
\description{
prep_bin_data
}
