% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_codeData.R
\name{prep_codeData}
\alias{prep_codeData}
\title{prep_codeData}
\usage{
prep_codeData(subData, subDataBin = NA, subDataMed = NA, refTx = NA)
}
\arguments{
\item{subData}{}

\item{subDataBin}{}

\item{subDataMed}{}

\item{refTx}{}
}
\value{
List of data and statistics.
If no binary of median data empty sub-lists.
}
\description{
Create input data for NMA.
}
