% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_conts_data.R
\name{prep_conts_data}
\alias{prep_conts_data}
\title{prep_conts_data}
\usage{
prep_conts_data(contsData, refTx = NA, covars = NA)
}
\arguments{
\item{contsData}{Counts data}

\item{refTx}{Reference treatment}
}
\value{

}
\description{
prep_conts_data
}
