% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_count_data.R
\name{prep_count_data}
\alias{prep_count_data}
\title{Prep Count Data}
\usage{
prep_count_data(count_dat, refTx, preRefTx = NA)
}
\arguments{
\item{count_dat}{}

\item{refTx}{}
}
\value{

}
\description{
Prep Count Data
}
