% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankProbPlot.R
\name{rankProbPlot}
\alias{rankProbPlot}
\title{Rank probability plot}
\usage{
rankProbPlot(nma, res_bugs, folder = "output", label = "", save = FALSE, ...)
}
\arguments{
\item{save}{Logical}

\item{...}{Additional arguments}
}
\value{

}
\description{
Treatment Ranking Plot.
}
