% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{rinits}
\alias{rinits}
\title{Randomly generate MCMC initial values}
\usage{
rinits(nTx, nStudies, param_names)
}
\arguments{
\item{nTx}{Number of treatments}

\item{nStudies}{Number of studies}

\item{param_names}{Parameter names}
}
\value{
closure
}
\description{
Randomly generate MCMC initial values
}
