% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_bugs_files.R
\name{save_bugs_files}
\alias{save_bugs_files}
\title{save_bugs_files}
\usage{
save_bugs_files(labels, bugs_filename, folder = "output", fileSep = "/")
}
\arguments{
\item{labels}{}

\item{bugs_filename}{}

\item{folder}{}

\item{fileSep}{}
}
\value{

}
\description{
save_bugs_files
}
