% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{setupData}
\alias{setupData}
\title{Set-up study data}
\usage{
setupData(
  subData,
  refTx = NA,
  subDataBin = NA,
  subDataMed = NA,
  is_random = TRUE
)
}
\arguments{
\item{subData}{Main data. Mandatory}

\item{refTx}{Reference treatment name}

\item{subDataBin}{Binary data. Optional}

\item{subDataMed}{Median time data. Optional}

\item{is_random}{Is this a random effects model? Logical}
}
\value{
list
}
\description{
Arrange input data for NMA.
}
