% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subData}
\alias{subData}
\title{Study mean and standard error treatment difference}
\format{
A data frame with 6 variables:
\describe{
\item{study}{Study name}
\item{base}{Reference treatment name}
\item{tx}{Comparison treatment name}
\item{Lmean}{Mean difference}
\item{Lse}{Standard error of difference}
\item{multi_arm}{Multi-arm trial indicator: 0/1}
}
}
\usage{
subData
}
\description{
A dataset containing the pair comparisons for studies.
}
\keyword{datasets}
