% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survDataBin}
\alias{survDataBin}
\title{Study survival binary outcome data}
\format{
A data frame with 5 variables:
\describe{
\item{study}{Study name}
\item{base}{Reference treatment name}
\item{tx}{Comparison treatment name}
\item{BinR}{Count of number of events}
\item{BinN}{Total sample size}
}
}
\usage{
survDataBin
}
\description{
A dataset containing the pair comparisons for studies.
}
\keyword{datasets}
