% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subDataHR}
\alias{subDataHR}
\title{Study log hazard ratios and standard errors}
\format{
A data frame with 6 variables:
\describe{
\item{study}{Study name}
\item{base}{Reference treatment name}
\item{tx}{Comparison treatment name}
\item{Lmean}{Log hazard ratio}
\item{Lse}{Standard error of log hazard ratio}
\item{multi_arm}{Multi-arm trial indicator: 0/1}
}
}
\usage{
subDataHR
}
\description{
A dataset containing the pair comparisons for studies.
Woods NMA allows data on HR on the log scale.
}
\keyword{datasets}
