% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subDataMed}
\alias{subDataMed}
\title{Study median time to event}
\format{
A data frame with 6 variables:
\describe{
\item{study}{Study name}
\item{base}{Reference treatment name}
\item{tx}{Comparison treatment name}
\item{median}{Median time}
\item{medN}{Sample size}
\item{medR}{Sample size/2 as estimated number of events}
}
}
\usage{
subDataMed
}
\description{
A dataset containing the pair comparisons for studies.
}
\keyword{datasets}
