% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surroNMA.R
\name{surroNMA-internal}
\alias{surroNMA-internal}
\title{Internal surroNMA functions}
\description{
Internal functions for surroNMA package. These functions are not intended
for direct use by users.
}
\keyword{internal}
