% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txEffectPlot.R
\name{txEffectPlot}
\alias{txEffectPlot}
\title{Treatment effect plot}
\usage{
txEffectPlot(
  dat,
  res_bugs,
  labels,
  endpoint = NULL,
  preRefTx = NA,
  save = FALSE,
  ...
)
}
\arguments{
\item{dat}{Study input data; list}

\item{res_bugs}{Simulations}

\item{labels}{Labels}

\item{endpoint}{End point name; string}

\item{preRefTx}{Reference treatment}

\item{...}{}
}
\value{
dat
}
\description{
Treatment effect plot
}
