% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{utils-surroNMA}
\alias{utils-surroNMA}
\alias{simulate_surro_data}
\alias{rank_from_draws}
\alias{sucra}
\alias{poth}
\alias{mid_adjusted_preference}
\title{Utility Functions}
\usage{
simulate_surro_data(
  K = 5,
  J = 20,
  per_study = 1,
  alpha = 0,
  beta = 0.8,
  tauS = 0.2,
  tauT = 0.3,
  sigma_d = 0.4,
  classes = NULL,
  seed = 1
)

rank_from_draws(eff_matrix)

sucra(ranks)

poth(ranks)

mid_adjusted_preference(eff_matrix, MID = 0)
}
\arguments{
\item{K}{Number of treatments}

\item{J}{Number of studies}

\item{per_study}{Studies per treatment}

\item{seed}{Random seed}
}
\value{
Various
}
\description{
Utility and helper functions
}
\keyword{internal}
