% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varPlot.R
\name{varPlot}
\alias{varPlot}
\title{Study network diagram}
\usage{
varPlot(
  xVar,
  yVar,
  showYLab = TRUE,
  showXLab = TRUE,
  pointText = NA,
  joinY = TRUE,
  txList = NA,
  folder,
  label,
  cex.axis = 0.2,
  ...
)
}
\arguments{
\item{xVar}{x variable}

\item{yVar}{y variable}

\item{showYLab}{logical}

\item{showXLab}{logical}

\item{pointText}{point text}

\item{joinY}{join y}

\item{txList}{List of treatments}

\item{folder}{Folder name; string}

\item{label}{label}

\item{cex.axis}{size axis}

\item{...}{Additional arguments}
}
\description{
Study network diagram
}
