\docType{data}
\name{woods1}
\alias{woods1}
\title{Network meta-analysis dataset of COPD: Hazard ratio statistics}
\format{A data frame for network meta-analysis with 2 trials.}
\usage{
data(woods1)
}
\description{
A network meta-analysis dataset for COPD summarized in hazard ratio statistics provided in Woods et al. (2010).
\itemize{
  \item \code{studlab}: ID variable of studies.
  \item \code{treat1}: Treatment 1.
  \item \code{treat2}: Treatment 2.
  \item \code{TE}: Log hazard ratio estimate.
  \item \code{seTE}: Standard error estimate of the log hazard ratio estimator.
  \item \code{n1}: Sample size 1.
  \item \code{n2}: Sample size 2.
}
}
\references{
Woods, B. S., Hawkins, N. and Scott, D. A. (2010). 
Network meta-analysis on the log-hazard scale, combining count and hazard ratio statistics accounting for multi-arm trials: A tutorial. 
\emph{BMC Medical Research Methodology} 
\strong{10}: 54.
}

\keyword{datasets}
