\docType{data}
\name{woods2}
\alias{woods2}
\title{Network meta-analysis dataset of COPD: Dichotomized data}
\format{A data frame for network meta-analysis with 3 trials.}
\usage{
data(woods2)
}
\description{
A network meta-analysis dataset for COPD reported only as dichotomized data provided in Woods et al. (2010).
\itemize{
  \item \code{study}: ID variable of studies.
  \item \code{treat}: Treatment.
  \item \code{d}: The number of events.
  \item \code{n}: Sample size.
}
}
\references{
Woods, B. S., Hawkins, N. and Scott, D. A. (2010). 
Network meta-analysis on the log-hazard scale, combining count and hazard ratio statistics accounting for multi-arm trials: A tutorial. 
\emph{BMC Medical Research Methodology} 
\strong{10}: 54.
}

\keyword{datasets}
