% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results_table.R
\name{write_results_table}
\alias{write_results_table}
\title{write_results_table}
\usage{
write_results_table(nma, res_bugs, folder = "output", label = "", ...)
}
\arguments{
\item{...}{}
}
\description{
write_results_table
}
