% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs_stats.R
\name{bugs_stats}
\alias{bugs_stats}
\title{bugs_stats}
\usage{
bugs_stats(dat, res_bugs, effectParam, random)
}
\arguments{
\item{dat}{input data; list}

\item{res_bugs}{output from running bugs}

\item{effectParam}{text vector of names}

\item{random}{RE or FE}
}
\value{

}
\description{
bugs_stats
}
