% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolders.R
\name{createFolders}
\alias{createFolders}
\title{Create output folders}
\usage{
createFolders(folder = "output", ...)
}
\arguments{
\item{folder}{Output folder. Text string}

\item{...}{Additional arguments}
}
\value{

}
\description{
Create output folders
}
