% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolders.R
\name{createFolders}
\alias{createFolders}
\title{Create output folders}
\usage{
createFolders(folder = "output", fileSep = "/", ...)
}
\arguments{
\item{folder}{Text string name}

\item{fileSep}{File separate; default forward slash}

\item{...}{Additional arguments}
}
\value{

}
\description{
Create output folders
}
