% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{WinBugs diagnostic plots}
\usage{
diagnostic_plots(
  bugs_filename,
  bugsData,
  bugs_params,
  bugs_fn,
  inits,
  params_to_save,
  winDebug = PAUSE,
  folder
)
}
\arguments{
\item{bugs_filename}{}

\item{bugsData}{}

\item{inits}{}

\item{params_to_save}{}

\item{winDebug}{}

\item{folder}{}
}
\description{
WinBugs diagnostic plots
}
