% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{WinBugs diagnostic plots}
\usage{
diagnostic_plots(res_bugs, labels, folder = "output", fileSep = "/")
}
\arguments{
\item{res_bugs}{}

\item{labels}{}

\item{fileSep}{}
}
\description{
WinBugs diagnostic plots
}
