% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{WinBUGS diagnostic plots}
\usage{
diagnostic_plots(res_bugs, labels_short = "", folder = "output")
}
\arguments{
\item{res_bugs}{Output from a BUGS run}

\item{folder}{Text string name}

\item{labels}{$short is used in file name save}

\item{fileSep}{File separator; default forward slash}
}
\description{
WinBUGS diagnostic plots
}
