% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bugs_filename.R
\name{make_bugs_filename}
\alias{make_bugs_filename}
\title{Make BUGS model file name}
\usage{
make_bugs_filename(random, data_type)
}
\arguments{
\item{random}{Type of model, RE or FE; logical}

\item{data_type}{Data type}
}
\value{
string
}
\description{
Make BUGS model file name
}
