% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bugs_filename.R
\name{make_bugs_filename}
\alias{make_bugs_filename}
\title{make_bugs_filename}
\usage{
make_bugs_filename(random, dat)
}
\arguments{
\item{random}{logical}

\item{dat}{list}
}
\value{
string
}
\description{
make_bugs_filename
}
