% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Make Report}
\usage{
make_report(nma_model, ext = "pdf", echo = FALSE, ...)
}
\arguments{
\item{ext}{A string of text to indicate the extension of the
resulting output file. Possible options are \code{"pdf"}, \code{"docx"}.
This requires the use of pandoc, knitr and rmarkdown.}

\item{echo}{A string (default to \code{FALSE}) to instruct whether
the report should also include the \code{BCEA} commands used to
produce the analyses. If the optional argument \code{echo} is set
to \code{TRUE} (default = \code{FALSE}), then the commands are also
printed.}

\item{...}{Additional parameters.}
}
\description{
Constructs the automated report from the output of the NMA.
}
\examples{

\dontrun{
}

}
\references{

}
