% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_bugs_summary.R
\name{my_bugs_summary}
\alias{my_bugs_summary}
\title{my_bugs_summary}
\usage{
my_bugs_summary(bugs_summary, params, txList)
}
\arguments{
\item{bugs_summary}{}

\item{params}{}

\item{colEff}{}
}
\value{

}
\description{
Function to set up results file for
multiple parameters monitored
}
