% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_NMA.R
\name{new_NMA}
\alias{new_NMA}
\title{NMA constructor}
\usage{
new_NMA(
  survDataHR = NA,
  survDataMed = NA,
  survDataBin = NA,
  binData = NA,
  countData = NA,
  contsData = NA,
  bugs_params = NA,
  is_random = TRUE,
  data_type = NA,
  hyperparams = list(),
  refTx = NA,
  effectParam = NA,
  label,
  endpoint
)
}
\arguments{
\item{survDataHR}{Hazard ratio input data frame. Optional}

\item{survDataMed}{Median time input data frame. Optional}

\item{survDataBin}{Survival binary data input data frame. Optional}

\item{binData}{Binary data input data frame. Optional}

\item{countData}{Count data input data frame. Optional}

\item{contsData}{Continuous data input data frame. Optional}

\item{bugs_params}{List of BUGS parameters. Optional}

\item{is_random}{Random effects model? Logical}

\item{data_type}{Vector of names of data formats from
"hr_data", "surv_bin_data", "med_data", "bin_data",
"count_data", "conts_data"}

\item{hyperparams}{List of hyperparameters}

\item{refTx}{Reference treatment; string}

\item{effectParam}{Effect parameter}

\item{label}{Label}

\item{endpoint}{End point name; string}
}
\value{

}
\description{
Create an \code{nma} class object to use in an analysis.
All of the data types are optional but at least one needs to be
passed to the function. The \code{data_type} must be a subset of
the passed data sets. If none given then all of the supplied
data are assumed.
}
\seealso{
\code{\link{NMA_run}}, \code{\link{NMA_update}}
}
