% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_NMA.R
\name{new_NMA}
\alias{new_NMA}
\title{nma constructor}
\usage{
new_NMA(
  subData,
  subDataMed = NA,
  subDataBin = NA,
  bugs_params = NA,
  is_random = TRUE,
  hyperparams = list(),
  refTx = NA,
  effectParam,
  modelParams,
  label,
  endpoint
)
}
\arguments{
\item{subData}{}

\item{subDataMed}{}

\item{subDataBin}{}

\item{bugs_params}{}

\item{is_random}{}

\item{refTx}{}

\item{effectParam}{}

\item{modelParams}{}

\item{label}{}

\item{endpoint}{}
}
\value{

}
\description{
nma constructor
}
