% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA.R
\name{NMA}
\alias{NMA}
\title{NMA}
\usage{
NMA(
  winSource,
  dataFunc,
  effectParam = NA,
  modelParams = NA,
  folder,
  label,
  endpoint,
  binData,
  medData,
  refTx = NA,
  preRefTx = NA,
  decEff,
  random = FALSE,
  lg
)
}
\arguments{
\item{winSource}{}
}
\value{

}
\description{
NMA
}
