% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_run.R
\name{NMA_run}
\alias{NMA_run}
\alias{NMA_run.nma}
\title{Run network meta-analysis}
\usage{
NMA_run(nma, output_dir = "output", save = TRUE)

\method{NMA_run}{nma}(nma, output_dir = "output", save = TRUE)
}
\arguments{
\item{nma}{Object of class \code{nma}}

\item{output_dir}{Output folder name; string}

\item{save}{Save BUGS output to file? Logical}
}
\value{
\code{res_bugs}


}
\description{
Run network meta-analysis
}
\seealso{
\code{\link{new_NMA}}, \code{\link{NMA_update}}
}
