% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_update.R
\name{NMA_update}
\alias{NMA_update}
\alias{NMA_update.nma}
\title{Update NMA}
\usage{
NMA_update(nma, ...)

\method{NMA_update}{nma}(nma, ...)
}
\arguments{
\item{nma}{Object of class \code{nma}}

\item{...}{Additional arguments}
}
\value{



}
\description{
Take an existing analysis and modify it.
}
\seealso{
\code{\link{new_NMA}}, \code{\link{NMA_run}}
}
