% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseTable.R
\name{pairwiseTable}
\alias{pairwiseTable}
\title{Pairwise table}
\usage{
pairwiseTable(
  nma,
  res_bugs,
  folder = "output",
  probs = c(0.025, 0.975),
  label = "",
  save = FALSE,
  ...
)
}
\arguments{
\item{nma}{Object of class \code{nma}}

\item{res_bugs}{Output from running BUGS model}

\item{folder}{Folder name; string}

\item{probs}{Vector of probabilities}

\item{label}{string}

\item{save}{Logical}

\item{...}{Additional arguments}
}
\value{
table
}
\description{
Pairwise table
}
