% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\alias{plotNetwork.default}
\alias{plotNetwork.nma}
\title{Plot network diagram}
\usage{
plotNetwork(dat, usecurve = FALSE)

\method{plotNetwork}{default}(dat, usecurve = FALSE, ...)

\method{plotNetwork}{nma}(dat, usecurve = FALSE, ...)
}
\arguments{
\item{dat}{List of study data, including subData
and possibly survDataBin and survDataMed}

\item{usecurve}{Which line to use? Logical}

\item{...}{Additional arguments}
}
\value{

}
\description{
Plot network diagram
}
