% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_and_tables.R, R/plots_and_tables_tmp.R
\name{plots_and_tables}
\alias{plots_and_tables}
\title{Write to file multiple NMA plots and tables}
\usage{
plots_and_tables(
  dat,
  res_bugs,
  effectParam,
  labels,
  endpoint = NULL,
  folder = "output",
  fileSep = "/"
)

plots_and_tables(
  dat,
  res_bugs,
  effectParam,
  labels,
  endpoint = NULL,
  folder = "output",
  fileSep = "/"
)
}
\arguments{
\item{dat}{List of study data, including subData
and possibly subDataBin and subDataMed}

\item{res_bugs}{BUGS output}

\item{effectParam}{Effect parameter names; string}

\item{labels}{Labels}

\item{endpoint}{End point names; string}

\item{folder}{Output folder name; string}

\item{fileSep}{File separator; default forward slash}
}
\value{
res_bugs

res_bugs
}
\description{
Write to file multiple NMA plots and tables

Write to file multiple NMA plots and tables
}
