% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_codeData.R
\name{prep_codeData}
\alias{prep_codeData}
\title{Prepare data used in computation}
\usage{
prep_codeData(
  subDataHR = NA,
  subDataBin = NA,
  subDataMed = NA,
  binData = NA,
  countData = NA,
  contsData = NA,
  refTx = NA
)
}
\arguments{
\item{subDataHR}{Hazard ratio data frame}

\item{subDataBin}{Optional data frame. Survival binary data}

\item{subDataMed}{Optional data frame. Median times}

\item{refTx}{Reference treatment name; string}
}
\value{
List of data and statistics.
If no binary of median data empty sub-lists.
}
\description{
Create input data for NMA.
Renames the study, base and other fields with appended letter
depending on type of data.
}
