% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_codeData.R
\name{prep_codeData}
\alias{prep_codeData}
\title{Prepare data used in computation}
\usage{
prep_codeData(
  survDataHR = NA,
  survDataBin = NA,
  survDataMed = NA,
  binData = NA,
  countData = NA,
  contsData = NA,
  refTx = NA
)
}
\arguments{
\item{survDataHR}{Hazard ratio data frame}

\item{survDataBin}{Optional data frame. Survival binary data}

\item{survDataMed}{Optional data frame. Median times}

\item{refTx}{Reference treatment name; string}
}
\value{
List of data and statistics.
If no binary of median data empty sub-lists.
}
\description{
Create input data for NMA.
Renames the study, base and other fields with appended letter
depending on type of data.
}
