% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankProbPlot.R
\name{rankProbPlot}
\alias{rankProbPlot}
\title{Rank probability plot}
\usage{
rankProbPlot(res_bugs, labels, dat, save = FALSE, ...)
}
\arguments{
\item{labels}{Labels}

\item{...}{}
}
\value{

}
\description{
Treatment Ranking Plot.
}
