% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankProbPlot.R
\name{rankProbPlot}
\alias{rankProbPlot}
\title{Rank probability plot}
\usage{
rankProbPlot(sims, labels)
}
\arguments{
\item{sims}{sims.matrix bugs output}

\item{labels}{}
}
\value{

}
\description{
Treatment Ranking Plot.
}
