% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankProbPlot.R
\name{rankProbPlot}
\alias{rankProbPlot}
\title{Rank probability plot}
\usage{
rankProbPlot(nma, res_bugs, folder = "output", label = "", save = FALSE, ...)
}
\arguments{
\item{nma}{Object of class \code{nma}}

\item{res_bugs}{Output from running BUGS model}

\item{folder}{Folder name; string}

\item{label}{string}

\item{save}{Logical}

\item{...}{Additional arguments}
}
\value{

}
\description{
Treatment Ranking Plot.
}
