% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_bugs_files.R
\name{save_bugs_files}
\alias{save_bugs_files}
\title{save_bugs_files}
\usage{
save_bugs_files(
  res_bugs,
  bugs_params,
  run_bugs,
  labels,
  folder = "output",
  fileSep = "/"
)
}
\arguments{
\item{res_bugs}{}

\item{bugs_params}{}

\item{run_bugs}{}

\item{labels}{}

\item{folder}{}

\item{fileSep}{}
}
\value{

}
\description{
save_bugs_files
}
