% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{setupData}
\alias{setupData}
\title{setupData}
\usage{
setupData(
  subData,
  refTx = NA,
  subDataBin = NA,
  subDataMed = NA,
  is_random = TRUE
)
}
\arguments{
\item{subData}{}

\item{refTx}{Reference treatment name}

\item{subDataBin}{}

\item{subDataMed}{}

\item{is_random}{Is this a random effects model?}
}
\value{
list
}
\description{
Arrange input data for NMA.
}
