% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{setupData}
\alias{setupData}
\title{Set-up study data}
\usage{
setupData(nma_datasets, data_type = NA, refTx = NA, is_random = TRUE)
}
\arguments{
\item{nma_datasets}{survDataHR: Hazard ratio data. Optional;
survDataBin: Survival binary data. Optional;
survDataMed: Median time data. Optional;
binData: Binary data. Optional;
countData: Count data. Optional;
contsData: Continuous data. Optional}

\item{data_type}{Data type}

\item{refTx}{Reference treatment name}

\item{is_random}{Is this a random effects model? Logical}
}
\value{
list
}
\description{
Arrange input data for NMA.
}
