% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txEffectPlot.R
\name{txEffectPlot}
\alias{txEffectPlot}
\title{Treatment effect plot}
\usage{
txEffectPlot(
  nma,
  res_bugs,
  label = "",
  folder = "output",
  endpoint = NULL,
  preRefTx = NA,
  save = FALSE,
  ...
)
}
\arguments{
\item{nma}{Study input data; list}

\item{res_bugs}{Simulations}

\item{label}{Label}

\item{folder}{String}

\item{endpoint}{End point name; string}

\item{preRefTx}{Reference treatment}

\item{save}{Logical}

\item{...}{Additional arguments}
}
\value{
dat
}
\description{
Forest plot of 2.5, median and 07.5\% quantiles of
posterior distributions for each treatment.
}
