% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txEffectPlot.R
\name{txEffectPlot}
\alias{txEffectPlot}
\title{Treatment effect plot}
\usage{
txEffectPlot(dat, sims, labels, endpoint = NULL, preRefTx = NA)
}
\arguments{
\item{dat}{Study input data; list}

\item{sims}{Simulations}

\item{labels}{}

\item{endpoint}{End point name; string}

\item{preRefTx}{Reference treatment}
}
\value{
dat
}
\description{
Treatment effect plot
}
