% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txEffectPlot.R
\name{txEffectPlot}
\alias{txEffectPlot}
\title{txEffectPlot}
\usage{
txEffectPlot(dat, sims, labels, preRefTx, refTx)
}
\arguments{
\item{dat}{input data list}

\item{sims}{}

\item{labels}{}

\item{preRefTx}{}

\item{refTx}{reference treatment}
}
\value{

}
\description{
txEffectPlot
}
