% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varPlot.R
\name{varPlot}
\alias{varPlot}
\title{Study network diagram}
\usage{
varPlot(
  xVar,
  yVar,
  showYLab = TRUE,
  showXLab = TRUE,
  pointText = NA,
  joinY = TRUE,
  txList = NA,
  folder,
  label,
  cex.axis = 0.2,
  ...
)
}
\description{
Study network diagram
}
