% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_to_file.R
\name{write_data_to_file}
\alias{write_data_to_file}
\alias{write_data_to_file.nma}
\title{Write study data to file}
\usage{
write_data_to_file(nma, label = "", folder = "output", ...)

\method{write_data_to_file}{nma}(nma, label = "", folder = "output", ...)
}
\arguments{
\item{nma}{Result of \code{new_NMA}}

\item{...}{Additional arguments}
}
\description{
Write study data to file
}
