% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results_table.R
\name{write_results_table}
\alias{write_results_table}
\title{Write results table}
\usage{
write_results_table(nma, res_bugs, folder = "output", label = "", ...)
}
\arguments{
\item{nma}{Object of class \code{nma}}

\item{...}{Additional arguments}
}
\description{
Write results table
}
